refactor(app): consolidate panel visibility helper
