--- packages/app/src/pages/session/side-panel-visibility.ts
@@ -1,3 +1,0 @@
-export function shouldShowFileTree(input: { desktopV2: boolean; showFileTree: boolean; opened: boolean }) {
-  return input.opened && (!input.desktopV2 || input.showFileTree)
-}
